대쉬보드용 API 추가 Event / Monitoring / Device Status

+++ src/main/java/egovframework/customize/service/ControlLogicDeviceVO.java
@@ -21,7 +21,7 @@ import lombok.Setter;
 @Setter
 public class ControlLogicDeviceVO {
 	Integer id;
-	Integer logicId;
+	Long logicId;
 	Integer deviceNum;
 	String deviceParamName;
 	String deviceType;
